msp430-4{5-7}: collapse package directory, move from all to ARCH
- all packages go in flat directory (no need to seperate)
- move msp430-45 (and others) from all to be architecture dependent.

The all debian archives generated for the msp430-4{5-7} packages
are different when built on an x86 (32 bit) vs. amd64 (64 bit) machines.
The checksums differ.   Don't know why but not worth messing with so
now seperated based on architecture.

Also now no point to seperating the resultant packages into different
directories.   So flattening.